Patches to enable 8173 EVB for ATF 0314-2016 (1.2+)
Add RESET_TO_BL31 supports, MTK internal preloader (which is
used to load ATF on EVB) is quite different from ATF BL2. And
the boot flow is tedious because the bootloader (BL33) (lk) is an
AArch32 one. See [1] for more details of the boot flow.

Current patch is a bit hacky. Will move to [1]-based later.

[1] https://github.com/ARM-software/arm-trusted-firmware/pull/506

Note that, with this patch ATF compiles, but the system won't
go to next stage correctly